[WIP] SQL; Factory->Lorhondel callback; Player repository freshen via HTTP GET; JSON validation always returns object
Freshen is currently non-functional in that the HttpServer receives the request but the 'client-side' repository's Browser does not receive the response nor throw an error for an invalid response.

Worst case scenario is abandoning processing repository saving within the Browser callback by handling that processing 100% within the web API, but this means that there can never be a true client-side application communicating with the web server as there will be no way to receive information. This method of transferring data to and from the HttpServer had been thoroughly tested up to the point of implementing this test method into the repository class as a function, so it is still unclear to me where the breakdown is happening.