Player activate and deactivate
The activate code is written in a way that assumes that there is an unknown number of activated players even though we know there should only ever be one. This is a performant failsafe that will deactivate any number of active players. While this could have been done without the promise string evaluation it would not have been possible to reliably retrieve and process the $result callback from the server. Because we want to be able to verify the callback, this method is required.

x2