Party repository method reworks; Misc changes to naming and data handling within methods; Docs
This change is mostly to allow for calling the function as a method within the party part, but also to accommodate calling the function from the repository.

isPartyFull is now also being aliased within the Party repository so the internal API can reference it in a cleaner way rather than grabbing the actual Party part and calling it. The reasoning behind this change is that the API may need to reference the function at this level in the future. This method will be removed upon future review if it is determined to not be necessary for the library.